test_task_backend/requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
numpy==1.20.3
pandas==1.2.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
pandas==1.2.4
